did a lot... chagned model, made restaurant recommendations page, made route for editing user based on kevins food preferences, refactored duplicate code, uncamelcased components
