+++ client/components/google-map.js
@@ -1,0 +1,30 @@
+import React, {Component} from 'react'
+import {withGoogleMap, GoogleMap, Marker} from 'react-google-maps'
+const {
+  MarkerWithLabel
+} = require('react-google-maps/lib/components/addons/MarkerWithLabel')
+
+const Map = props => {
+  const {longitude, latitude} = props
+  console.log(props)
+  const GoogleMapExample = withGoogleMap(props => (
+    <GoogleMap defaultCenter={{lat: latitude, lng: longitude}} defaultZoom={13}>
+      <MarkerWithLabel
+        position={{lat: latitude, lng: longitude}}
+        labelAnchor={new google.maps.Point(0, 0)}
+        labelStyle={{}}
+      >
+        <div />
+      </MarkerWithLabel>
+    </GoogleMap>
+  ))
+  return (
+    <div>
+      <GoogleMapExample
+        containerElement={<div style={{height: `200px`, width: '600px'}} />}
+        mapElement={<div style={{height: `100%`}} />}
+      />
+    </div>
+  )
+}
+export default Map
