+++ client/components/index.js
@@ -9,4 +9,4 @@ export {default as MergedAuth} from './merged-auth'
 export {default as Homepage} from './homepage'
 export {default as Restaurant} from './restaurant-list'
 export {default as Options} from './options'
-export {default as Profile} from './profile'
+export {default as Recommendations} from './recommendations'
