+++ client/components/restaurant-list.js
@@ -16,6 +16,7 @@ import {
   Loader,
   Dimmer
 } from 'semantic-ui-react'
+import {restaurantSort, restaurantScorer} from '../helperFuncs'
 
 class RestaurantList extends Component {
   constructor() {
@@ -41,8 +42,9 @@ class RestaurantList extends Component {
       foursquareWeight,
       restaurants
     } = this.props
+
     restaurants.forEach(restaurant => {
-      restaurant.score = this.restaurantScore(
+      restaurant.score = restaurantScorer(
         restaurant.reviews,
         yelpWeight,
         zomatoWeight,
@@ -70,57 +72,6 @@ class RestaurantList extends Component {
     )
 
     this.setState({loading: false})
-  }
-
-  restaurantScore = (
-    reviews,
-    yelpWeight,
-    zomatoWeight,
-    googleWeight,
-    foursquareWeight
-  ) => {
-    let totalWeight = yelpWeight
-    let reviewsString = ''
-    reviews.forEach(review => {
-      reviewsString += review.source
-    })
-    const reviewsLength = reviews.length
-
-    if (reviewsString.indexOf('Zomato') !== -1) totalWeight += zomatoWeight
-    if (reviewsString.indexOf('Google') !== -1) totalWeight += googleWeight
-    if (reviewsString.indexOf('Foursquare') !== -1)
-      totalWeight += foursquareWeight
-
-    function weighter(sourceWeight, review) {
-      const weight = sourceWeight / totalWeight * reviewsLength
-      let denominator
-      if (review.source === 'Foursquare') {
-        denominator = 10
-      } else {
-        denominator = 5
-      }
-      const rating = review.rating / denominator
-      return weight * rating
-    }
-
-    return Math.round(
-      reviews
-        .map(review => {
-          if (review.source === 'Yelp') {
-            return weighter(yelpWeight, review)
-          } else if (review.source === 'Zomato') {
-            return weighter(zomatoWeight, review)
-          } else if (review.source === 'Google') {
-            return weighter(googleWeight, review)
-          } else if (review.source === 'Foursquare') {
-            return weighter(foursquareWeight, review)
-          }
-        })
-        .reduce((accum, currentVal) => accum + currentVal, 0) /
-        reviews.length *
-        100,
-      0
-    )
   }
 
   async handleChange(event) {
@@ -187,7 +138,7 @@ class RestaurantList extends Component {
     } = this.props
 
     this.props.restaurants.forEach(restaurant => {
-      restaurant.score = this.restaurantScore(
+      restaurant.score = restaurantScorer(
         restaurant.reviews,
         yelpWeight,
         zomatoWeight,
@@ -230,23 +181,11 @@ class RestaurantList extends Component {
         />
         <Divider hidden />
         <Card.Group>
-          {restaurants
-            .sort((restaurant1, restaurant2) => {
-              if (restaurant2.reviews.length === restaurant1.reviews.length) {
-                if (restaurant2.score === restaurant1.score) {
-                  return restaurant2.id - restaurant1.id
-                } else {
-                  return restaurant2.score - restaurant1.score
-                }
-              } else {
-                return restaurant2.reviews.length - restaurant1.reviews.length
-              }
-            })
-            .map(restaurant => {
-              return (
-                <RestaurantCard restaurant={restaurant} key={restaurant.id} />
-              )
-            })}
+          {restaurantSort(restaurants).map(restaurant => {
+            return (
+              <RestaurantCard restaurant={restaurant} key={restaurant.id} />
+            )
+          })}
         </Card.Group>
       </React.Fragment>
     )
