+++ server/db/models/user.js
@@ -16,6 +16,9 @@ const User = db.define('user', {
       return () => this.getDataValue('password')
     }
   },
+  likedCuisines: {
+    type: Sequelize.ARRAY(Sequelize.TEXT)
+  },
   salt: {
     type: Sequelize.STRING,
     // Making `.salt` act like a function hides it when serializing to JSON.
@@ -26,7 +29,7 @@ const User = db.define('user', {
   },
   googleId: {
     type: Sequelize.STRING
-  },
+  }
 })
 
 module.exports = User
